// src/app/apiv2/model/userResponseDTO.ts
/**
 * Bestellsystem API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponseDTO { 
    id?: string;
    surname?: string;
    name?: string;
    email?: string;
}